% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_boostedRIRT.R
\name{boostedRIRT}
\alias{boostedRIRT}
\title{Erzeuge ein boostedRIRT-Objekt}
\usage{
boostedRIRT(X, Y, id, colsample, subsample, alpha, iter_final, maxdepth = 3)
}
\description{
Erzeuge ein boostedRIRT-Objekt
}
\keyword{internal}
